+++ lib/workbook-parser.js
@@ -34,25 +34,23 @@ function findSheetsToSave(wb) {
 
 function parseWorkbook(wb) {
     const sheetsToSave = findSheetsToSave(wb);
-    for (let s of sheetsToSave) {
-        if (!wb.Sheets[s]) throw Error(`Sheet [${s}] not found`);
-    }
+    const firstMissingSheet = sheetsToSave.find(s => !wb.Sheets[s]);
+    if (firstMissingSheet) throw Error(`Sheet [${firstMissingSheet}] not found`);
 
     const result = {};
-    const formatters = {
-        date: (d) => {
-            if (!d) return '';
-            const day = d.getDate().toString().padStart(2, '0');
-            const month = (1 + d.getMonth()).toString().padStart(2, '0');
-            const year = d.getFullYear().toString();
-            return day + '.' + month + '.' + year;
-        },
-        // number: (name, formatted) => formatted,
-    };
     for (let s of sheetsToSave) {
         try {
             result[s.toLowerCase()] = sheetToJson(wb.Sheets[s], {
-                formatters,
+                formatters: {
+                    date: (d) => {
+                        if (!d) return '';
+                        const day   = d.getDate().toString().padStart(2, '0');
+                        const month = (1 + d.getMonth()).toString().padStart(2, '0');
+                        const year  = d.getFullYear().toString();
+                        return day + '.' + month + '.' + year;
+                    },
+                    // number: (name, formatted) => formatted,
+                },
                 keepEmptyRows: true,
                 firstRowCommentChar: '#',
                 trimOnEmptyHeader: true,
